% Auto-generated: do not edit by hand
\name{dashWeather}

\alias{dashWeather}

\title{DashWeather component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashWeather(id=NULL, api_key=NULL, city=NULL, lang=NULL, lat=NULL,
lon=NULL, service=NULL, style=NULL, unit=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{api_key}{Character. ----.}

\item{city}{Character. ----.}

\item{lang}{Character. ----.}

\item{lat}{Numeric. ----.}

\item{lon}{Numeric. ----.}

\item{service}{Character. ----.}

\item{style}{Named list. ----.}

\item{unit}{Character. ----.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

